fix(desktop): guard the quick ask home URL, and move the region list out of the component file
React Doctor flagged two warnings on the custom cloud work. The quick ask
window now checks the dev server URL with URL.canParse before it parses it,
and getSelectableRegions moves to its own module so RegionSelect exports
components only.

Generated-By: PostHog Desktop
Task-Id: 78590351-3721-441c-9dd2-73329784b38e

Generated-By: PostHog Desktop
Task-Id: 78590351-3721-441c-9dd2-73329784b38e
