+++ products/desktop/apps/code/src/main/quick-ask.ts
@@ -272,14 +272,18 @@ function createQuickAskWindow(): BrowserWindow {
   // it gets the same navigation boundary as the main window: links open in
   // the external browser and the window itself never leaves its own page —
   // an in-place navigation would carry the bridges into a foreign origin.
-  const quickAskHome = QUICK_ASK_VITE_DEV_SERVER_URL
-    ? new URL(`${QUICK_ASK_VITE_DEV_SERVER_URL}/quick-ask.html`)
-    : pathToFileURL(
-        path.join(
-          __dirname,
-          `../renderer/${QUICK_ASK_VITE_NAME}/quick-ask.html`,
-        ),
-      );
+  const quickAskUrl = QUICK_ASK_VITE_DEV_SERVER_URL
+    ? `${QUICK_ASK_VITE_DEV_SERVER_URL}/quick-ask.html`
+    : null;
+  const quickAskHome =
+    quickAskUrl && URL.canParse(quickAskUrl)
+      ? new URL(quickAskUrl)
+      : pathToFileURL(
+          path.join(
+            __dirname,
+            `../renderer/${QUICK_ASK_VITE_NAME}/quick-ask.html`,
+          ),
+        );
   setupExternalLinkHandlers(window, quickAskHome);
 
   window.setAlwaysOnTop(true, "screen-saver");
